rename util to log
