+++ src/raft/util.go
@@ -1,78 +1,22 @@
 package raft
 
-import (
-	"fmt"
-	"log"
-	"os"
-	"strconv"
-	"time"
-)
-
-var debugStart time.Time
-var debugVerbosity int
-
-func init() {
-	debugVerbosity = getVerbosity()
-	debugStart = time.Now()
-
-	log.SetFlags(log.Flags() &^ (log.Ldate | log.Ltime))
-}
-
-type logTopic string
-
-const (
-	dClient  logTopic = "CLNT"
-	dCommit  logTopic = "CMIT"
-	dDrop    logTopic = "DROP"
-	dError   logTopic = "ERRO"
-	dInfo    logTopic = "INFO"
-	dLeader  logTopic = "LEAD"
-	dLog     logTopic = "LOG1"
-	dLog2    logTopic = "LOG2"
-	dPersist logTopic = "PERS"
-	dSnap    logTopic = "SNAP"
-	dTerm    logTopic = "TERM"
-	dTest    logTopic = "TEST"
-	dTimer   logTopic = "TIMR"
-	dTrace   logTopic = "TRCE"
-	dVote    logTopic = "VOTE"
-	dWarn    logTopic = "WARN"
-)
-
-// Retrieve the verbosity level from an environment variable
-func getVerbosity() int {
-	v := os.Getenv("VERBOSE")
-	level := 0
-	if v != "" {
-		var err error
-		level, err = strconv.Atoi(v)
-		if err != nil {
-			log.Fatalf("Invalid verbosity %v", v)
+// lastLogIdxWithTerm use binary search to find the last entry that has the same term as input.
+// if there is none, return false.
+func lastLogIdxWithTerm(log []LogEntry, term int) (bool, int) {
+	l, r := 0, len(log)-1
+	for l < r {
+		mid := (l + r) / 2
+		if log[mid].Term > term {
+			r = mid - 1
+		} else if log[mid].Term < term {
+			l = mid + 1
+		} else {
+			for log[mid].Term == term && mid < len(log)-1 {
+				mid++
+			}
+			return true, mid - 1
 		}
 	}
-	return level
-}
 
-func Debug(topic logTopic, format string, a ...interface{}) {
-	if debugVerbosity >= 1 {
-		time := time.Since(debugStart).Microseconds()
-		time /= 100
-		prefix := fmt.Sprintf("%06d %v ", time, string(topic))
-		format = prefix + format
-		log.Printf(format, a...)
-	}
-}
-
-func Max(a, b int) int {
-	if a > b {
-		return a
-	}
-	return b
-}
-
-func Min(a, b int) int {
-	if a < b {
-		return a
-	}
-	return b
+	return false, -1
 }
